move craftassist_specs to lowlevel
