+++ droidlet/lowlevel/minecraft/craftassist_specs.py
@@ -6,7 +6,7 @@ import numpy as np
 import os
 import pickle
 
-PATH = os.path.join(os.path.dirname(__file__), "lowlevel/minecraft/minecraft_specs")
+PATH = os.path.join(os.path.dirname(__file__), "minecraft_specs")
 assert os.path.isdir(PATH), (
     "Not found: "
     + PATH
